If given a SHA use that to test.
If we are passed a SHA we should use that to test that any previously
downloaded file is correct and not just rely on its existence.

Also, set an unless on the SHA exec so that it doesnt execute everytime if
the SHA is correct causing other resources to be possibly notified.